Add SignUp and SignIn components
