--- src/features/auth/sign-in/components/Logo/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Text, View } from 'react-native';
-
-const Logo = () => (
-  <View>
-    <Text>My Custom Logo</Text>
-  </View>
-);
-
-export default Logo;
