--- src/features/auth/sign-in/screen/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import { AuthLayout } from 'features/auth/layout';
-
-import Logo from '../components/Logo';
-import { SignInPropTypes } from './types';
-
-const SignInScreen: React.FunctionComponent<SignInPropTypes> = () => (
-  <AuthLayout>
-    <Logo />
-  </AuthLayout>
-);
-
-export default SignInScreen;
