--- src/features/auth/sign-in/screen/types.ts
@@ -1,3 +1,0 @@
-import { AuthStackParamList, NativeStackScreenProps } from 'navigation/types';
-
-export type SignInPropTypes = NativeStackScreenProps<AuthStackParamList, 'SignIn'>;
